openvibe-plugins-tools :  * several changes for windows portability
git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@359 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe-plugins/tools/trunc/src/ovpCEBMLStreamSpy.cpp
@@ -138,8 +138,8 @@ void CEBMLStreamSpy::openChild(const EBML::CIdentifier& rIdentifier)
 	{
 		cout << "  ";
 	}
-	uint32 l_ui32Identifier1=(((EBML::uint64)rIdentifier)>>32);
-	uint32 l_ui32Identifier2=(((EBML::uint64)rIdentifier)&0xffffffff);
+	uint32 l_ui32Identifier1=(uint32)(((EBML::uint64)rIdentifier)>>32);
+	uint32 l_ui32Identifier2=(uint32)(((EBML::uint64)rIdentifier)&0xffffffff);
 	cout
 		<< "Opened EBML node [id:"
 		<< ConsoleColor_FG_Blue
